mmc: sdio: Fix invalid vdd in voltage switch power cycle
[ Upstream commit d9bfbb95ed598a09cf336adb0f190ee0ff802f0d ]

The 'ocr' parameter passed to mmc_set_signal_voltage()
defines the power-on voltage used when power cycling
after a failure to set the voltage.  However, in the
case of mmc_sdio_init_card(), the value passed has the
R4_18V_PRESENT flag set which is not valid for power-on
and results in an invalid vdd.  Fix by passing the card's
ocr value which does not have the flag.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Cc: stable@vger.kernel.org # v3.13+
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Sasha Levin <sasha.levin@oracle.com>
Git-commit: e6cbcb008a279d7794fba01799ecfb6be391484f
Git-repo: https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git
Change-Id: I1d49eed013cabd918e18450593fd4a7dcdf03ef9
Signed-off-by: Veerabhadrarao Badiganti <vbadigan@codeaurora.org>
Signed-off-by: Ryan Andri <ryanandri@linuxmail.org>